Use testdriver Action API in WPT pointerevent tests - Part 6
In order to run the web-platform-tests automatically, we will use
testdriver Action API in all the wpt to simulate inputs. Here we are
changing some pointerevent input tests which have touch, mouse and pen
inputs and complex actions.

This is the sixth CL that changes part of the pointerevent input tests.

Bug: 606367
Change-Id: I1e4a273fed440ba23f851a11f6e8c3eac8d18c77